requirements.txt
pyyaml
pyogg
pyopenal

